fix: compatibility with plugins that use strict-version-matcher-plugin
The build of a mobile app with both the OneSignal
and Firebase Mobile plugins fails. The OneSignal
gradle script changes the version of all
‘com.google.android.gms’ dependencies to ‘11.8+’,
including a Firebase Mobile dependency
(['strict-version-matcher-plugin']) that doesn’t have
the ‘11.8+’ version.

This commit fixes that by adding said plugin to
a list of plugins where the version override logic
is not applied to.

References https://outsystemsrd.atlassian.net/browse/RNMT-3264